TSAL Control (READ DESC.)
TSAL Control board, moved here from AMU-V00. Current iteration is being sent for manufacture, hence this file should not need to be editted unless there was an error in design.
Note the Schematic/PCB libraries in this folder (currently TSAL Control) are not, at the time of writing, the same libraries as the ones outside the folder, and should not be mixed until they have been updated together.